rename sql_utils to utils
